new script + improvements

+++ codeSize.sh
@@ -5,8 +5,10 @@ mkdir -p "results"
 repo='git@github.com:SebsLittleHelpers/scala-js.git'
 repoFolder='repo'
 
-rm -rf "$repoFolder"
-git clone "$repo" "$repoFolder"
+if [ -d "$repoFolder" ]; then
+  rm -rf "$repoFolder"
+  git clone "$repo" "$repoFolder"
+fi
 
 cd $repoFolder
 
@@ -24,7 +26,7 @@ for branch in "master" "isInstanceOf"; do
   sbt 'reversi/fullOptJS'
   mv "examples/reversi/target/scala-2.11/reversi-opt.js" "../results/$branch-reversi-fullopt.js"
   
-  # TODO replace with command to generate js file, not actually run it
+
   sbt 'testSuite/test:fastOptJS'
   mv "test-suite/js/target/scala-2.11/scalajs-test-suite-test-fastopt.js" "../results/$branch-test-fastopt.js"
